+++ libft/libft.h
@@ -40,7 +40,7 @@ char				*ft_strchr(const char *str, int ch);
 char				*ft_strrchr(const char *str, int ch);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
 char				*ft_strnstr(const char *big, const char *lit, size_t len);
-long				ft_atoi(char **str);
+int 				ft_atoi(char **str);
 int					ft_isalpha(int ch);
 int					ft_isdigit(int ch);
 int					ft_isalnum(int ch);
